pull out sort function so it can be used on multiple events
